🏗️ Refactor: 미디어 검색 fetch 커스텀 훅으로 관리
